requirements.txt
asgiref==3.8.1
boto3==1.35.82
botocore==1.35.82
Brotli==1.1.0
dj-database-url==2.3.0
Django==5.1.4
django-debug-toolbar==4.4.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
